+++ test/unit/test_registry_get_request_timeout.py
@@ -5,5 +5,5 @@ from dbt.clients.registry import _get
 
 class testRegistryGetRequestTimeout(unittest.TestCase):
     def test_registry_request_timeout(self):
-        # using non routable IP to test timeout logic in the _get function
-        self.assertRaises(Timeout, _get, '', 'https://getdbt.com:81')
+        # using closed port to test timeout logic in the _get function
+        self.assertRaises(Timeout, _get, '', 'https://www.getdbt.com:81')